requirements.txt
psutil
timeago
matplotlib
discord.py
aiosqlite
jishaku
polib
Unidecode
wargaming
Pillow
pycryptodomex
tqdm
mock
ddt
raven
lxml
imageio
imageio-ffmpeg
numpy